Slicing changes
Slicing had a bug in labeling edges in the backward slicing case.
The edge from the instruction at which backward slicing
is applied, was labeled with the output of the instruction. Instead we
should create an edge for each of the inputs to the instruction (since
it is backward slicing) and label them based on the input.