Force FP saves
As per user request, we are adding a force FP save option.
This will force FP registers to be saved. This is necessary because
right now, even if saveFP is set, analysis may incorrectly determine
no FP is ised and hence not save FP registers. This happens if
the instrumentation calls a function using FP in the rewriter case.
forceFPSave interface will not be published - since we hope to fix this for
the next release.