powerpc rewriter commit
This commit implements more accurate heuristics for parsing
powerpc jump tables using backward dataflow graph analysis.

This commit also implements heuristics and backward slicing
to determine if a blr instruction in a powerpc binary,
is a return or a jump table. This is specially relevant to BlueGene binaries.
To do slicing, an instruction adapter is always constructed/associated
with a specific block.