Tries system directory for windows libraries that have no path
When creating a mapped file, we need a path for it, but we don't
always get one when a library is loaded (e.g., for kernel32.dll and
ntdll.dll).  We now have the mapped path deduce the path of the
dll by reading the "windir" environment variable which is set by
the OS itself, and tack that on to dll's that have no path and can't
be found without one.