Removes unsafe cache usage on stopThreads and post-call-patch fix
We were using the cache for indirect calls that could leave the module,
but we can't do that if we're also emulating memory accesses, since
we may need to synchronize shadow and orig memory around these calls.
Also, our post-call patching incorrectly assumed that the instrumentation
for the currently executing function occurred in the most recent round of
relocations.