Implement handling for fork/exec/exit syscalls, and thread creation and exit.
Also, updates the Makefile for x86 Linux to use the new internal process
control layer. Fixes a bug in code gen for iRPCs where an incorrect result
register is returned by AST code generation, which results in the iRPC handling
code grabbing the result of the RPC from the wrong register.

Includes some refactoring of the BPatch layer to simplify the handling of
system calls and also destruction of BPatch_process and BPatch_thread objects.

Fixed a race condition with waiting for processes to stop.
BPatch_process::isTerminated should not poll for events and before waiting for
events in BPatch::waitForStatusChange, we should check and make sure some
processes are running as a sanity check.