Overwrite, synchronization, and shared code fixes
  Drew's consistency checks discovered that overwrites were failing
when shared code was deleted, I fixed this by not finalizing
functions that have had shared blocks removed from them until
the block has been removed from all functions (in the interim the
block is still in the CFG and was getting added back in by finalization).
  I also cleaned up a check in AddressSpace::deleteSnippet that was
overly aggressive and belonged in the hybridAnalysis code, this was
causing us not to remove loop instrumentation when we'd detected
overlapping overwrite loops.
  Shadow copies were failing when we jumped into another module rather
than calling into it, since we didn't know where to put the post-call
orig-to-shadow copy.  Fixed this by dropping the instrumentation after
the jump's callers.
  Shadow copies were also failing due to overlapping springboards, just
removing an assert was not enough, since we would copy a negative amount.