Fix a few ProcControlAPI bugs.
1) When a process is forked, refresh_libraries should be called so
the underlying AddressTranslate objects can be created for the new
process. This also means that the mem_state fork constructor should
not copy over libraries because they will be correctly created when
refresh_libraries is called.

2) Certain internal continues on Linux for a thread should not continue
the thread with a pending signal. This case occurs when a signal is
received after an event that is currently being handled and the
handling for this event requires an internal continue.