Adding new platform:ppc32_bgp
This commit adds a new platform - ppc32_bgp. For now,
ppc32_bgp is the same as ppc32_linux for components.
This will change when proccontrol is commited.

The only difference between ppc32_linux and ppc32_bgp is
in the testsuite. For ppc32_bgp, we build only rewriter mutatees
and run them on on the compuet nodes. This requires a new
environment variable called "DYNINST_BGP_PARTITION" to be set.
This env variable determines which BGP partition the rewritten
mutatees must be executed.