A few bug fixes/enhancements for ProcControlAPI
Adds a new event, ForceTerminate. A callback for this event is delivered
when a process exits after a user requests a process be terminated. It's
parent event is of type Terminate.

Fixes some bugs related to stopping a thread that is in the process of
exiting (i.e., after pre-Exit but before post-Exit). There is no guarantee
that an exiting thread will be delivered a pending stop, and as such, there
are a few corner cases that require different behavior if the thread is
exiting.

Added parameters to Process::createProcess and related methods to allow the
environment of the new process to be specified.