sigHandler handling, hideDebugger, shared lib instrum., rdtsc parse output
The faultInsnAddr in the EXCEPTION_RECORD for int3's is one less
than the PC at which the program resumes execution, and I wasn't
taking that into account.  There was also a bug in the code that
patches in a different address at which to resume execution after
the exception.
 HideDebugger wasn't patching BlockInput properly, since we wind
up having to relocate the function to figure out where it returns
to, and were therefore executing the unpatched version.
 The parser now produces output when it hits instructions that
might be used to detect us: like rdtsc and sldt.
 Skirts problems related to not being able to re-instate write
protections for a page in Yoda's Protector that's accessed by
kernel32.dll 0x7c864143 by handling the violation and restoring
write permissions before the page executes, whenever this occurs
for an instruction in a non-defensive library, as it shouldn't
actually be overwriting code.