Fix for bug in loading RT library during attach on Linux x86/x86_64.
To load the RT library into a process Dyninst has attached to, we
generate some code to load the library via libdl or to an internal libc
function, if libdl isn't in the process. The generated code was not
following the ABI and allowing space for a function to store local
variables below the stack pointer. Our base tramp code already takes
this possibility into account so the solution was to wrap this
generated code with the standard RPC header and trailer that internally
do a base tramp save and restore, respectively.