More fixes for ProcControlAPI
Handling for a few cases of sending stops to exiting threads was missing.  The
first is the pre-exit of the initial thread. It wasn't being marked as exiting
and as such, ProcControl was waiting for pending stops that would never arrive.

The second case is where the stop and the pre-exit occur simultaneously (at
least from ProcControl's perspective). The generator processes the stop before
the pre-exit and as a result, when processing a pre-exit, an assertion was
failing because the thread was expected to be running instead of stopped. The
handling for this case requires setting a flag in the thread when the generator
sees an exit event so the assertion can check for this case.

Finally, when a pre-exit is received before a SIGSTOP (that may never come),
the pre-exit handler needs to mark the process stopped to maintain consistency
between the actual state of the process and the state maintained by
ProcControl.

Fixes a bug with proc stopper events (such as a Breakpoint) where a event
would get dropped on the floor when the handler thread attempted to handle
a proc stopper event with a registered callback. Since the event has
a registered callback, the handler cannot handle the event. At this point,
the event is not added back to the collection of proc stopper events. The
solution is to only remove the event from the collection when it can be
handled.