Fix a bug with LD_PRELOAD, add some debugging support code, small changes to test_driver.
Modified the test_driver to do program_teardown after all the tests for a
specific module (aka Dyninst, Symtab, etc) have been run instead of when all
the tests for all modules have been executed. In the Dyninst component's
teardown, the BPatch object is destructed which in turn makes sure the
callback thread exits. If this doesn't occur before ProcControl tests are
executed, the callback thread will interfere with the event handling
performed by the user thread in the ProcControl tests, causing them to
exhibit confusing behavior.

The LD_PRELOAD cannot be set in the Dyninst parent process when creating a
mutatee because this setting will persist for multiple mutatees. This causes
problems for when working with 32-bit mutatees.

Adds some debugging support code to dump stack walks when a process recieves a
crash signal and debugging output is on.

Modifies the Dyninst break point on Linux to use tkill. It turns out that
there is no guarantee that kill(lwp, signal) will deliver the signal to the
correct lwp. To get this guarantee, we use tkill.

Modified the implementation of the tc_lock_lock in the RT library on Linux.
There was a possible ordering of threads that would result in the lock failing.
I didn't directly observe this ordering, but it could in theory cause problems.