Instrumentation of replaced WinAPIs, post-call parsing
Keeps a map in the HybridAnalysis that tracks replaced functions so
we know to instrument the replacement instead of the original,
which has been blown away.

Fixes post-call parsing, it parsing post-call instructions as
new functions if there was a tail-call in-between the caller
and callee; this was causing problems with post-call padding
never being taken care of.