Implementation of control-transfer breakpoints
A control transfer breakpoint is a breakpoint where after the
breakpoint is hit, we change the PC to a pre-defined address. This
required adding some code to the HandlePostBreakpoint event handler
to set the PC to this pre-defined address when handling a control
transfer breakpoint.

Also includes a small fix for cases when the execv fails on Unix
platforms. The child fails the exec and exits. At exit, the ProcControl
library teardown code kicks in and was deadlocking on the ProcPool
lock. The solution was to unlock the ProcPool lock in the child
before the exec.