Trap-based instrumentation using control transfer breakpoints for dynamic instrumentation.
Adds some hooks to AddressSpace to allow the PCProcess to insert
control transfer breakpoints when a trap mapping is added. These hooks
don't do anything in BinaryEdit at the moment.

Also adds some handling for control-transfer breakpoints; the
corresponding events don't require user interaction so the ProcControl
event for these breakpoints is handled by the callback thread and
not passed to the Dyninst user.