Parsing performance improvement.
The destructor for an IA_IAPI object at local scope inside the parsing
loop was determined to be responsible for 5-9% of total parsing time.

By replacing the IA_IAPI internal stl::map with an stl::vector, we are
able to move the object outside the loop, and reset it in constant time.

Thanks to Nate for finding and solving this problem.