Enabling DyninstAPI with ProcControlAPi on power Linux
Refactored RegisterConversion to allow a uniform interface for
converting Dyninst Register representations to ProcControlAPI's
MachRegister representations.

Refactored RT library load AST creation to linux.C as most of it
is not architecture specific (the only remaining architecture
specific functionality is in createUnprotectStackAST).

Removed syscall tracing as it is no longer required to locate main. The
syscall tracing was used to determine when libc was loaded. When this
occurs, we could determine the location of main from libc, if we can't
determine its address from our other heuristics. ProcControlAPI gives
us events when libc is loaded so we don't need to trace system calls
to get this event.

Removed vsyscall indentification from Dyninst -- this is now provided
by StackwalkerAPI.

Removed some old process control code from StackWalker to get it building
on Power Linux. This code is no longer needed with the integration of
ProcControlAPI into StackWalker.