test5_6 fix
Based on Nate's suggestion:
test 5_6 is broken as currently implemented, given what the parser is
capable of doing to resolve catch blocks---a correctly functioning
parser can't find the catch blocks that test 5_6 seeks to test. One
obvious fix is to make func_that_throws (actually
exception_test::call_cpp) have a path to return, even if it is never
taken (think volatile int a; a = 0; if(!a) { sample_exception(); }
else { return; }). The only other way to fix this would be to rework
the way exception catch blocks are located, which reworking may not
even be feasible.