Fix for force loading RT library via an iRPC
When libc and ld.so are stripped, our heuristics for identifying
a dlopen-like function to load the RT library were failing.

This commit adds functionality to use the __libc_dlopen_mode
function when libc is stripped. This function is exported via the
dynamic symbol table so it will be there even when the library
is stripped.

This approach is only used when symbols are not available because
it relies on an internal libc flag (__RTLD_DLOPEN) for correct
functionality on x86. See the comments in the code for more info.