Fix for determining if the process has passed main on ppc64 Linux.
To determine if the process is passed main, we determine the
entry addr of the dynamic linker and compare it to the pc of
the process. In the create case, these two values match indicating
that the process hasn't passed main.

On ppc64 Linux, the entry addr is actually the address of the opd
for the first function executed by the dynamic linker. This commit
adds code to dereference the entry addr, if necessary.