Enable CPU and memory usage tracking in parseThat and the testsuite.
The -memcpu flag for the testsuite no longer requires a parameter.
Additionally, CPU (user + system) and memory (RSS or maxRSS if available)
is reported for each test regardless of output driver.

The -memcpu flag is also passed to parseThat based tests.  In theory, this
should be a more reliable source for historic CPU and memory usage data.