Fix for force loading the RT library on Linux when libc and the loader are stripped.
Adds a new heuristic to process::loadDYNINSTlib() on both power and x86
that searches for the function __libc_dlopen_mode and uses this function
to function to force load the RT library. According to the comments in
the glibc source, __libc_dlopen_mode is used by libraries that libc
internally dlopen's via its internal mechanisms. Its interface is
the same as the exported dlopen call in libdl.

This heuristic is applied after searching for the exported version of
dlopen and before searching for do_dlopen, which requires that the
loader have symbols.

On x86, we need to add another mode flag that is not available in a
public header to turn off a stack protection check that is usually done
by the exported dlopen function before callling libc's internal dlopen
machinery. When symbols are available, we don't rely on this special
flag (see tryUnprotectStack for more info).