ppc64 OPD offset fix
In ppc64 platform, functions symbols (like _init, _fini etc)
are stored in .opd (Official Procedure Descriptor) section.
We have to find the ELF section where the function actually exists
and set the region for the function correctly. For example, _init will be
in .init section and not .opd section. There was a bug in determining is an
offset exists in an ELF section. This is fixed now.

This bug caused the _fini symbol to point to a wrong ELF section and
hence init_fini_callback test fails on ppc64 (bug 1157).