Fixes to thread_db support for FreeBSD
The most recent set of changes to the thread_db integration did not
work on FreeBSD. This commit fixes these bugs.

The biggest change is that FreeBSD needs to use PT_RESUME and PT_SUSPEND
to change the running state of threads once a process is stopped.

Also includes some changes to the thread test to account for a
difference in thread events and information available on FreeBSD.