Temporary fix for ppc32_linux failures
Previous commit (7237b6688017860a02a3d9254000e3665f5490c0) added a sort
routine to sort vector of edges. This caused a segfault on ppc32_linux
as the pointers used to sort may be NULL. This is a temporary fix for the
segfault. This fix ensures determinism of the order of the edges. A better
fix may follow in the future.