ProcControlAPI single-stepping support for ppc Linux
There were five sets of changes that go into providing single-stepping
support on ppc Linux. Single-stepping on ppc Linux requires being able
to detect atomic instruction sequences and use breakpoints to
effectively single step these instructions as a group.

1) Added an emulated single-step internal feature to support single
stepping with breakpoints. This required a new class emulated_singlestep
that represents a collection of int_breakpoints that are added to and
removed from a process as a whole. An emulated_singlestep can include
multiple breakpoints because the atomic instruction sequences can
contain control transfers. This also involved significant changes to
the breakpoint handlers to remove a emulated_singlestep after it
has been hit.

2) Bug fixes for when multiple threads hit a breakpoint before it is
handled. Before, a single step breakpoint would be removed in a
handler and a following Breakpoint event would go to work with the
breakpoint and run into problems. The fix was to store threads that
are currently clearing an installed_breakpoint with the
installed_breakpoint. Then, when removing the breakpoint, it is not
uninstalled until all threads and int_breakpoints no longer are
using the installed_breakpoint.

3) Bug fixes for reading registers from a pre-exited thread. This
causes an oops on ppc Linux and the kernel kills the thread that
attempted to read the registers. The workaround was to avoid
reading registers on a pre-exited thread on ppc Linux.

4) Added field to int_thread to store a pc before a single step on
ppc Linux. This field is needed to distinguish a single step to
an instruction that is a breakpoint and a single step that triggers
a breakpoint. In both cases, the PC after the corresponding trap
is the same.

5) Added code specific to ppc Linux to detect atomic instruction
sequences. This detection occurs immediately before attempting to
single step a thread. If a sequence is detected, the insertion of
the breakpoint is deferred until a process stop if other threads
are currently running. This deferred-process stop is triggered
with an asynchronous event PrepSingleStep.