Initial support for dynamic instrumentation on FreeBSD.
This support is non-functional right now, it is a WIP.

Includes a fair amount of refactoring to share code between
Linux and FreeBSD.

Refactored syscallNotification to be platform-independent except
for names of system call functions.

Small fix for ProcControlAPI on FreeBSD where a thread wasn't being
resumed after all other threads in a process had already exited.