Fix for generating a call using insnCodeGen::generatePush64
We were incorrectly doing a cast of a 16-bit value to an unsigned char
and then assigning it to a short in generatePush64. This was causing
code generated to jump to trampolines to be incorrect.

Also includes some fixes specific to FreeBSD.

Improves some error handling and reporting in inferiorMalloc and
attaching to a non-existent process.