Fixes for force loading the RT library on FreeBSD
Also, includes refactoring of our use of realpath. Our existing
use was non-portable in that we relied on realpath to also check
for the existence of the last element in the path. Linux performs
this check while FreeBSD does not.

All references to realpath replaced with a call to resolve_file_path
available in common/src/pathName.C.

Added the ability to disassemble jumps and multitramps insert into
the process using InstructionAPI. This helps with debugging our
use of instruction sequences that confuse objdump and friends. The
environment variable DYNINST_DEBUG_DISASS turns on this debugging
output.

Fixes for address translation that removes the non-portable
assumption that the link map doesn't contain an entry for the
executable. Linux, no executable. FreeBSD, executable is there.