Fix to enable instrumentation at main in dynamic mode
A breakpoint is set at the entry to main to determine when
the RT library can be loaded and system call instrumentation
can be inserted into the program.

ProcControlAPI's default treatment of breakpoints is to suspend
the breakpoint, single step and resume after the Breakpoint event
callback. Dyninst would like the process to stay at the entry
of main after hitting the breakpoint to enable function entry
point instrumentation at main.

The solution is to remove the breakpoint in the ProcControlAPI
callback -- ProcControl will not single step the process, keeping
the process at the entry point of main.