parse-after-call, block splitting, Win trap handling & buggy import rewrite
- fixes parseAPI.vcproj's post-compile directory creation & copying
- re-organizes parse-after-call and eliminates duplicate edge creation
- block splitting fixed for the non-defensive case
- gets trap handling working for binary rewriting of windows binaries
- adds buggy rewriting of the import table, and tries to force
libdyninstAPI_RT.dll to be added, though unsuccessfully:

BUG 1: Import rewriting that uses maps and therefore sorts the IAT,
which is bad, you can't change the order of existing IAT entries,
or function calls will go to the wrong targets.
BUG 2: You can't force a library to be added by creating an import
descriptor table entry for it and leaving the import lookup & address
tables empty.
- need to see if adding two IDT entries for the same library is valid
- can we split the IAT into chunks?  If not, we need to add bootstrap
code to the program to copy the linked IAT to its original address.