Fix for code used to find _start on ELF platforms
In image::findMain(), we add a symbol for _start if it doesn't
exist. When main exists, _start was mistakenly marked as
non-existent and later code adds a symbol for it. Stackwalking uses
this symbol to find the bottom of the stack and when a new process'
executable has already been parsed, the wrong _start symbol is given
to stackwalking resulting in a failure to find the bottom of the
stack.

We no longer add a _start symbol if one already exists. This allows
stackwalking to find the bottom of the stack when a binary has
already been parsed.