Implement temporary detach and re-attach for ProcControlAPI.
This feature required a new state for int_process and int_thread:
detached. A temporary detach suspends all breakpoints, sets the
process state and state of threads to detached. Unlike normal detach,
the int_process is not cleaned up after the detach completes.

Control methods for Process and Thread have been modified to check
whether the process/thread is detached before performing the operation.
If it is detached, an error condition is returned and the last error
is set to err_detached.

A reattach is similar to an attach except for the following differences:
* since all int_threads already exist, a different path is needed in the
  post attach handling to make sure the threads are issued low-level
  attaches. int_process::post_attach now takes a boolean that
  specifies whether we are doing an attach or re-attach.
* We need to issue thread destruction events for threads that no longer
  exist in the process
* The breakpoints that were suspended need to be resumed (i.e. written
  back into memory).

This commit also includes a new ProcControlAPI test: pc_temp_detach.
This test basically includes a mutatee that triggers some periodic
events. The mutator temporarily detaches from the mutatee, makes
sure it doesn't get these events. It then re-attaches to the
mutatee and reads some state out of the mutatee to verify that the
events did occur.