Rewriter RTlib load & Windows, Win-liveness, malware stats, vcproj fix
1. Rewriter fixes:
- we fail to load the runtime library in dynamic linking mode when we
  want tramp-guard support, but haven't added any traps to the program
- Some fixes to Windows rewriter so that it can access the trap table
  in the runtime library for trap-based instrumentation (doesn't work yet)
2. Windows liveness: Windows binaries that are built by Visual Studio
in debug mode add a call at function exits to ensure that the stack
was cleaned up properly, but this little check violates our assumption
that flags are not live across function calls, so we have to assume that
flags are live at function calls for Windows binaries.
3. We now output a lot of statistics about any binary that is analyzed
in defensive mode.
4. Added a missing file in the dyninstAPI.vcproj project file.