Commit: test1_37 fix
There were two problems with test1_37
1) The mutator was not checking if a variable is NULL.
Hence, the test crashes. This is fixed by adding an
check.
2) The mutatee uses static declaration for global
variables. bgxlc doesn't create a symbol with name
for static variables, since static variables
cannot be not accessed outside the file. This is a
fundamental problem that Dyninst cannot fix, if
the compiler decides to ignore the variable name.
The current fix is to make the global variable not
static.