Relocation information can be used to parse unlinked objects
Adds relocation entries to a hash map so that the relocations
can be looked up efficiently by address.  This allows the ParseAPI
allowing unlinked calls to look up call targets for which relocations
will patch in a different address, and keeps us from parsing at
placeholder values in call instructions.  Unfortunately, we're not
extracting the right relocation information yet, so this code doesn't
do much yet