This commit is to fix the bug due to some typos in PatchAPI:
In patchAPI/src/PatchMgr.C, wantXXX functions (e.g., wantFuncs, wantCalls ...)
use bitwise-or to test whether the types parameter passed from findPoints includes
a particular point type, which is incorrect. We should use bitwise-and to test.
So I replace all bitwise-or expression in wantXXX functions with the invocation
of a static function Point::TestType that does bitwise-and operation