BGQ bug fixes
This commit has two fixes:
1) Disk offset of new PT_LOAD segment in rewritten static binary
was set incorrectly. This is fixed now.

2) The maximum offset reachable using TOC pointer is 0x7fff (15 bits + 1 sign bit).
For static binaries, the TOC pointer must be able to reach the new load segment.
In Bluegene environment, we align by page size (1MB) but this causes the
the TOC pointer offset to be too large to reach the new segment. Since we
do not create a new PT_LOAD segment, but rather extend the existing PT_LOAD
segment, we do not need to align by 1MB. Instead, we align by minimal safe
number of bytes.