Fix --trace feature of parseThat.
The named pipe was being unlinked before the mutatee was instructed to
open the other end.  This also removes the attempt to access errno
inside the mutatee during tracepipe init errors.  It was causing
additional error messages that distract from the root cause.

Thanks to Romulo Curty Cerqueira for the bug report.