Elf_X: Fix the search for the build-id
The code was previously acting as though section .note.gnu.build-id
could be read directly as an ASCII string, but this is incorrect.  There
is at least an Elf_Nhdr first, and the actual build-id is in raw bytes.
It's also not guaranteed that the build-id will be in that named
section, as it could be merged with other SHT_NOTE sections.

This patch first adds an Elf_X_Nhdr type to mirror the other Elf_X
abstractions.  For the build-id it walks over all SHT_NOTE sections
looking for type==NT_GNU_BUILD_ID and name=="GNU", and then converts
that to a hex string for the /usr/lib/debug/.build-id/... path.