[API] Change sw_c out_frame parameters into actual outputs
The out_frame parameters of walkSingleFrame and getInitialFrame are
frame_t, a struct which is passed by value.  GCC 4.8 warns that this
parameter is set but not used:

Compiling ../src/sw_c.C
../src/sw_c.C: In function ‘int walkSingleFrame(walker_t, frame_t, frame_t)’:
../src/sw_c.C:117:5: warning: parameter ‘out_frame’ set but not used [-Wunused-but-set-parameter]
 int walkSingleFrame(walker_t walker, frame_t frame, frame_t out_frame)
     ^
../src/sw_c.C: In function ‘int getInitialFrame(walker_t, frame_t)’:
../src/sw_c.C:131:5: warning: parameter ‘out_frame’ set but not used [-Wunused-but-set-parameter]
 int getInitialFrame(walker_t walker, frame_t out_frame)
     ^

In order for this to be an actual output parameter, these will have to
be frame_t* parameters.  This is a simple change, but it does change the
API of these functions.  Given that they formerly couldn't work at all,
I expect no one is using these anyway...