Use a generic error message for gettid failure
In ProcSelf::getDefaultThread, it was using a temporary string to build
and error message with strerror, for setLastError.  This is an "escape",
because the c_str() value is invalid as soon as the string goes out of
scope.  But since gettid should never ever fail, and even if it did the
reason is logged by sw_printf, we can just be generic for setLastError.

Reported by Coverity as WRAPPER_ESCAPE (CWE-825).