[ABI] Return const std::string& for Statement::getFile()
This was returning a std::string, and BPatch_statement::fileName() was
then returning c_str() from that.  But since that string was temporary,
the memory backing it will be immediately freed.

The change to const std::string& shouldn't make any practical API
difference to callers, but it is an ABI change. (Though even that is
only if that simple function is ever not inlined...)

Reported by Coverity as WRAPPER_ESCAPE (CWE-825).