Assert that dwarf_loclist_from_expr_a returns cnt==1
Coverity reports ARRAY_VS_SINGLETON in DwarfWalker::decodeExpression,
because the address of a singleton, &descs, is passed with cnt to
decodeLocationListForStaticOffsetOrAddress, which treats it as an array.
Without knowledge of cnt, the analyzer thinks this could be source of
memory corruption.

In fact, dwarf_loclist_from_expr_a always sets listlen (cnt) to 1 on
success.  If we assert this, then the analyzer can tell that the
singleton "array" bound is fine.