proccontrol: Fix a copy-paste typo in ppc_process::readInsnForSS
After iterating over all of map mm_for_ss, this function will create a
new mem_response for the request.  If that reports isReady() right away,
then the memcpy incorrectly copies from the previous map iterator, which
is now exhausted (==end()).  This appears to be a copy-paste typo of the
memcpy from within the for loop, and this memcpy should use new_resp.

Reported by Coverity as INVALIDATE_ITERATOR.