dyninstAPI: Set image::findMain's findings as LOCAL/INTERNAL
When we look for new symbols in image::findMain, it's not really the
intent to make these global/exported symbols.  It doesn't matter so much
for live processes, but these shouldn't be exported in rewritten
binaries.  This patch sets them as LOCAL and INTERNAL instead.

Note, this is only for Linux and FreeBSD, but Windows may want it too.