dyninstAPI: Pad the fixed x86 instructions
The arrays used to generate illegal & trap instructions can be static
const, but still the analyzer doesn't see that the instruction decoder
won't actually read past their ends.  It's not a big deal to just pad
both illegalRep[] and trapRep[] to 8 bytes to be sure.

Reported by coverity as OVERRUN.